+++ components/elem-filter.tsx
@@ -5,13 +5,13 @@ import { convertToInternationalCurrencySystem } from "@/utils";
 import { getDefaultFilter, getFilterOptionMetadata } from "@/utils/filter";
 import { Filters, FilterOptionKeys, DateRangeOptions, DateCondition } from "@/models/Filter";
 import { ElemButton } from "./elem-button";
-import { InputRadio } from "./InputRadio";
-import { ElemTagsInput } from "./ElemTagsInput";
-import { ElemMultiRangeSlider } from "./ElemMultiRangeSlider";
+import { InputRadio } from "./input-radio";
+import { ElemTagsInput } from "./elem-tags-input";
+import { ElemMultiRangeSlider } from "./elem-multi-range-slider";
 import { InputDate } from "./input-date";
-import { ElemFilterPopup } from "./ElemFilterPopup";
-import { ElemAddFilter } from "./ElemAddFilter";
-import ElemAddressFilter from "./ElemAddressFilter";
+import { ElemFilterPopup } from "./elem-filter-popup";
+import { ElemAddFilter } from "./elem-add-filter";
+import ElemAddressFilter from "./elem-address-filter";
 import { InputText } from "./input-text";
 import { InputSelect } from "./input-select";
 import { eventSizeChoices } from "@/utils/constants";
