+++ components/event/elem-sponsor-grid.tsx
