+++ components/events/elem-event-card.tsx
