+++ components/events/elem-featured-events.tsx
@@ -2,14 +2,14 @@ import React, { FC } from "react";
 import { PlaceholderEventCard } from "@/components/placeholders";
 import { ElemCarouselWrap } from "@/components/elem-carousel-wrap";
 import { ElemCarouselCard } from "@/components/elem-carousel-card";
-import { useRouter } from "next/router";
 import {
 	Events_Bool_Exp,
-	Maybe,
+	GetEventsQuery,
 	Order_By,
 	useGetEventsQuery,
 } from "@/graphql/types";
-import { ElemEventCard } from "../Events-Temp/ElemEventCard";
+import { ElemEventCard } from "./elem-event-card";
+import moment from "moment-timezone";
 
 export type DeepPartial<T> = T extends object
 	? {
@@ -19,27 +19,24 @@ export type DeepPartial<T> = T extends object
 
 type Props = {
 	className?: string;
-	currentSlug: Maybe<string>;
-	tag1?: string;
-	tag2?: string;
+	heading?: string;
+	itemsLimit?: number;
 };
 
-export const ElemSimilarEvents: FC<Props> = ({
-	className,
-	currentSlug,
-	tag1,
-	tag2,
+export const ElemFeaturedEvents: FC<Props> = ({
+	className = "",
+	heading,
+	itemsLimit,
 }) => {
-	const limit = 12;
+	const limit = itemsLimit ? itemsLimit : 33;
 	const offset = null;
-
-	const router = useRouter();
 
 	const filters: DeepPartial<Events_Bool_Exp> = {
 		_and: [
 			{
-				slug: { _neq: "" || currentSlug },
-				_or: [{ types: { _contains: tag1 } }, { types: { _contains: tag2 } }],
+				slug: { _neq: "" },
+				is_featured: { _eq: true },
+				//end_date: { _gte: moment().subtract(1, "days").format("YYYY-MM-DD") },
 			},
 		],
 	};
@@ -51,32 +48,17 @@ export const ElemSimilarEvents: FC<Props> = ({
 	} = useGetEventsQuery({
 		offset,
 		limit,
-		order: Order_By.Desc,
+		order: Order_By.Asc,
 		where: filters as Events_Bool_Exp,
 	});
 
-	const events = eventsData?.events;
-
-	const onClickType = (
-		event: React.MouseEvent<HTMLDivElement>,
-		type: string
-	) => {
-		event.stopPropagation();
-		event.preventDefault();
-
-		router.push(
-			`/events/?filters=${encodeURIComponent(
-				`{"eventType":{"tags":["${type}"]}}`
-			)}`
-		);
-	};
+	const events: GetEventsQuery["events"] = eventsData?.events || [];
 
 	return (
-		<section className={`bg-white rounded-lg p-5 shadow ${className}`}>
-			<h2 className="text-xl font-bold">Similar Events</h2>
-
+		<div className={`bg-white rounded-lg p-5 ${className}`}>
+			{heading && <h2 className="text-xl font-bold">{heading}</h2>}
 			{error ? (
-				<h4>Error loading similar events</h4>
+				<h4>Error loading featured events</h4>
 			) : isLoading ? (
 				<>
 					<div className="flex overflow-hidden -mx-3">
@@ -93,19 +75,19 @@ export const ElemSimilarEvents: FC<Props> = ({
 			) : (
 				events && (
 					<ElemCarouselWrap>
-						{events.map((event: any) => {
+						{events.map((event) => {
 							return (
 								<ElemCarouselCard
 									key={event.id}
 									className={`p-3 basis-full sm:basis-1/2 lg:basis-1/3`}
 								>
-									<ElemEventCard event={event} onClickType={onClickType} />
+									<ElemEventCard event={event} />
 								</ElemCarouselCard>
 							);
 						})}
 					</ElemCarouselWrap>
 				)
 			)}
-		</section>
+		</div>
 	);
 };
