+++ components/my-list/people-list.tsx
@@ -11,7 +11,7 @@ import {
 	useGlobalFilter,
 } from "react-table";
 import { TableColumnsFilter } from "./table-columns-filter";
-import { TableGlobalFilter } from "./TableGlobalFilter";
+import { TableGlobalFilter } from "./table-global-filter";
 import { ElemPhoto } from "@/components/elem-photo";
 import {
 	IconSortUp,
