+++ components/person/elem-news-list.tsx
