+++ pages/events.tsx
@@ -2,7 +2,7 @@ import type { NextPage, GetStaticProps } from "next";
 import React, { Fragment, useState, useEffect } from "react";
 import toast, { Toaster } from "react-hot-toast";
 import { ElemHeading } from "../components/elem-heading";
-import { ElemFeaturedEvents } from "@/components/Events-Temp/ElemFeaturedEvents";
+import { ElemFeaturedEvents } from "@/components/events/elem-featured-events";
 import { ElemButton } from "../components/elem-button";
 import { runGraphQl } from "../utils";
 import { useStateParams } from "@/hooks/use-state-params";
@@ -20,10 +20,10 @@ import {
 import { DeepPartial } from "./companies";
 import { onTrackView } from "@/utils/track";
 import { useRouter } from "next/router";
-import { ElemFilter } from "@/components/ElemFilter";
+import { ElemFilter } from "@/components/elem-filter";
 import { processEventsFilters } from "@/utils/filter";
 import useFilterParams from "@/hooks/useFilterParams";
-import { ElemEventCard } from "@/components/Events-Temp/ElemEventCard";
+import { ElemEventCard } from "@/components/events/elem-event-card";
 import { useIntercom } from "react-use-intercom";
 
 type Props = {
